requirementsTF.txt
absl-py==0.15.0
astor==0.8.1
astunparse==1.6.3
cached-property==1.5.2
cachetools==4.2.4
certifi==2022.6.15
charset-normalizer==2.0.12
clang==5.0
cycler==0.11.0
dataclasses==0.8
decorator==4.4.2
facenet-pytorch==2.5.2
flatbuffers==1.12
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.47.0
h5py==3.1.0
idna==3.3
imageio==2.15.0
importlib-metadata==4.8.3
importlib-resources==5.4.0
keras==2.6.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.7
matplotlib==3.3.4
networkx==2.5.1
numpy==1.19.5
oauthlib==3.2.0
opencv-python==4.6.0.66
opt-einsum==3.3.0
Pillow==8.4.0
pkg_resources==0.0.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
PyWavelets==1.1.1
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.17.2
scipy==1.5.4
six==1.15.0
tensorboard==1.12.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.6.0
tensorflow-gpu==1.12.0
termcolor==1.1.0
tifffile==2020.9.3
torch==1.10.1
torchvision==0.11.2
tqdm==4.64.0
typing-extensions==3.7.4.3
urllib3==1.26.11
Werkzeug==2.0.3
wrapt==1.12.1
zipp==3.6.0
